New Component banner area
* Consolidates logic of what should be shown above the header in 1 place.
* Fixes bug CASH-1629
* Removes duplicate code between appeal banner and global promo banner
